Modified Search Bar
Given white background
